✨ feat: Implement UpdateAppointment API with `UpdateAppointmentCommand`
✅ Added `UpdateAppointmentCommand` to handle appointment updates.
🔄 Implemented `UpdateAppointmentCommandHandler` to process update logic.
📝 Endpoint: `PUT /api/appointments/{id}` for modifying existing appointments.
✅ Added validation to ensure updates meet business rules.
🧪 Included unit tests to verify command and handler functionality.

This feature allows users to update appointment details while maintaining data integrity and validation checks.
