+++ src/DoctorAppointmentService.Api/DTOs/Requests/UpdateAppointmentRequest.cs
@@ -1,7 +1,6 @@
 ﻿namespace DoctorAppointmentService.Api.DTOs.Requests;
-public class UpdateAppointmentDto
+public class UpdateAppointmentRequest
 {
-    public string Id { get; set; }
     public string PatientName { get; set; }
     public string DoctorId { get; set; }
     public DateTime AppointmentDate { get; set; }
